backend/requirements.txt
Flask
Flask-Limiter
requests
PyGithub
python-dotenv
semver
flask-cors
cachetools
